backend/requirements.txt
# Django 웹 프레임워크
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
uvicorn==0.34.3
gunicorn==23.0.0

# AI/LLM 관련
anthropic==0.51.0
openai==1.78.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.16
langchain-google-genai==2.1.4
langchain-upstage==0.7.1
langchain-chroma==0.2.4
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-sdk==0.1.66
langsmith==0.4.1

# 데이터베이스
chromadb==1.0.13
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
pymongo==4.13.2

# 유틸리티
python-dotenv==1.1.0
requests==2.32.3
pydantic==2.11.4
pydantic-settings==2.10.1
numpy
pandas
tiktoken==0.9.0
httpx==0.28.1
aiohttp==3.12.13
pillow==11.2.1